refactor: move mappers from service layer to api layer
